Align multiple ggplot graphs, taking into account the legend

This function aligns multiple ggplot graphs by making them the same width by taking into account the legend width.

AlignPlots(pltlist = NULL)

Arguments

pltlist
A list of plots (TODO: provide more detail)

Value

List of ggplot with equal width.

References

https://stackoverflow.com/questions/26159495/align-multiple-ggplot-graphs-with-and-without-legends

Examples

if (FALSE) {
library(ggplot2)
library(gtable)
library(cowplot)

## create 2 graphs
p1 <- ggplot2::ggplot(adtte, ggplot2::aes(x = as.numeric(AGE), fill = "Age")) +
  ggplot2::geom_histogram(bins = 15)
p2 <- ggplot2::ggplot(adtte, ggplot2::aes(x = as.numeric(AGE))) +
 ggplot2::geom_histogram(bins = 15)

## default alignment does not take into account legend size
cowplot::plot_grid(plotlist = list(p1,p2), align = "none", nrow=2)

## Alignplots takes into account legend width
cowplot::plot_grid(plotlist = AlignPlots(pltlist = list(p1, p2)), align = "none", nrow=2)
}