Summarize Hazard Ratio from a survival object using S3 method

S3 method for extracting information regarding Hazard Ratios. The function allows the survival object's formula to be updated. No default method is available at the moment.

get_COX_HR(x, ...)

# S3 method for survfit
get_COX_HR(x, update_formula = NULL, ...)

Arguments

x
An object of class survfit

...
other arguments passed on to the method survival::coxph

update_formula
Template which specifies how to update the formula of the survfit object update.formula

Value

A tidied object of class coxph containing Hazard Ratios

See also

Examples

## treatment effect
survfit_object_trt <- visR::estimate_KM(data = adtte, strata = c("TRTP"))
visR::get_COX_HR(survfit_object_trt)

#> tidyme S3 default method (broom::tidy) used.

#> # A tibble: 2 x 5
#>   term                     estimate std.error statistic  p.value
#>   <chr>                       <dbl>     <dbl>     <dbl>    <dbl>
#> 1 TRTPXanomeline High Dose     1.61     0.233      6.92 4.45e-12
#> 2 TRTPXanomeline Low Dose      1.42     0.230      6.20 5.71e-10

## treatment and gender effect
survfit_object_trt_sex <- visR::estimate_KM(data = adtte, strata = c("TRTP", "SEX"))
visR::get_COX_HR(survfit_object_trt_sex)

#> tidyme S3 default method (broom::tidy) used.

#> # A tibble: 3 x 5
#>   term                     estimate std.error statistic  p.value
#>   <chr>                       <dbl>     <dbl>     <dbl>    <dbl>
#> 1 TRTPXanomeline High Dose    1.62      0.234      6.91 4.96e-12
#> 2 TRTPXanomeline Low Dose     1.48      0.232      6.37 1.92e-10
#> 3 SEXM                        0.375     0.165      2.27 2.34e- 2

## update formula of KM estimates by treatment to include "SEX" for HR estimation
visR::get_COX_HR(survfit_object_trt, update_formula = ". ~ . + SEX")

#> tidyme S3 default method (broom::tidy) used.

#> # A tibble: 3 x 5
#>   term                     estimate std.error statistic  p.value
#>   <chr>                       <dbl>     <dbl>     <dbl>    <dbl>
#> 1 TRTPXanomeline High Dose    1.62      0.234      6.91 4.96e-12
#> 2 TRTPXanomeline Low Dose     1.48      0.232      6.37 1.92e-10
#> 3 SEXM                        0.375     0.165      2.27 2.34e- 2

## update formula of KM estimates by treatment to include "AGE" for
## HR estimation with ties considered via the efron method
visR::get_COX_HR(survfit_object_trt, 
  update_formula = ". ~ . + survival::strata(AGE)", ties = "efron")

#> Warning: Loglik converged before variable  3,7,18,36,37 ; coefficient may be infinite. 

#> tidyme S3 default method (broom::tidy) used.

#> # A tibble: 37 x 5
#>    term                        estimate std.error statistic  p.value
#>    <chr>                          <dbl>     <dbl>     <dbl>    <dbl>
#>  1 TRTPXanomeline High Dose       1.57      0.256   6.13    8.52e-10
#>  2 TRTPXanomeline Low Dose        1.32      0.253   5.22    1.76e- 7
#>  3 survival::strata(AGE)AGE=52  -18.2    5299.     -0.00343 9.97e- 1
#>  4 survival::strata(AGE)AGE=54   -0.209     1.42   -0.148   8.83e- 1
#>  5 survival::strata(AGE)AGE=56   -1.62      1.10   -1.48    1.40e- 1
#>  6 survival::strata(AGE)AGE=57   -1.14      1.17   -0.977   3.29e- 1
#>  7 survival::strata(AGE)AGE=59  -18.2    3768.     -0.00482 9.96e- 1
#>  8 survival::strata(AGE)AGE=60   -1.32      1.24   -1.07    2.86e- 1
#>  9 survival::strata(AGE)AGE=61   -1.45      1.13   -1.28    2.00e- 1
#> 10 survival::strata(AGE)AGE=62   -1.41      1.43   -0.986   3.24e- 1
#> # … with 27 more rows