Summarize the test for equality across strata from a survival object using S3 method

S3 method for extracting information regarding equality across strata. No default method is available at the moment.

get_pvalue(x, ...)

# S3 method for survfit
get_pvalue(
  survfit_object,
  ptype = "All",
  rho = NULL,
  statlist = c("test", "Chisq", "df", "p"),
  ...
)

Arguments

x
S3 object

...
other arguments passed on to the method

survfit_object
An object of class survfit

ptype
Character vector containing the type of p-value desired. Current options are "Log-Rank" "Wilcoxon" "Tarone-Ware" "Custom" "All". "Custom" allows the user to specify the weights on the Kaplan-Meier estimates using the argument rho. The default is "All" displaying all types possible. When rho is specified in context of "All", also a custom p-value is displayed.

rho
a scalar parameter that controls the type of test.

statlist
Character vector containing the desired information to be displayed. The order of the arguments determines the order in which they are displayed in the final result. Default is the test name ("test"), Chisquare test statistic ("Chisq"), degrees of freedom ("df") and p-value ("p").

Value

A tibble with summary measures for the Test of Equality Across Strata

See also

Author

Steven Haesendonckx

Examples

library(survival)
library(dplyr)
library(tidyr)
library(ggplot2)

survfit_object <- estimate_KM(data = adtte, strata = "TRTP")
get_pvalue(survfit_object)

#> # A tibble: 3 x 4
#>   `Equality across strata` Chisq    df `p-value`
#>   <chr>                    <dbl> <dbl> <chr>    
#> 1 Log-Rank                  60.3     2 <0.001   
#> 2 Wilcoxon                  48.0     2 <0.001   
#> 3 Tarone-Ware               41.9     2 <0.001