Extended quantile S3 method - EXPERIMENTAL NEEDS TESTING AND COMPARING WITH SAS SOFTWARE

S3 method for extracting quantiles. For survival objects, it takes into account the maximal time of survival as limit instead of NA If the quantile requested is larger than the minimal quantile available, the result is NA. No default method is available at the moment.

get_quantile(x, ...)

# S3 method for survfit
get_quantile(
  survfit_object = survfit_object,
  probs = c(0.25, 0.5, 0.75),
  conf.int = TRUE,
  tolerance = sqrt(.Machine$double.eps),
  debug = FALSE
)

Arguments

x
S3 object

...
other arguments passed on to the method

survfit_object
An object of class survfit

probs
probabilies Default = c(0.25,0.50,0.75)

conf.int
should lower and upper confidence limits be returned?

tolerance
tolerance for checking that the survival curve exactly equals one of the quantiles

debug
debug mode. Default = FALSE

Value

A tibble with quantiles

See also

Author

Steven Haesendonckx

Examples

# \donttest{
library(survival)
library(dplyr)
library(tidyr)

survfit_object_trt <- estimate_KM(data = adtte, strata = c("TRTP"))
get_quantile(survfit_object_trt)

#> # A tibble: 9 x 5
#>   strata                    quantity  `25`  `50`  `75`
#>   <chr>                     <chr>    <dbl> <dbl> <dbl>
#> 1 TRTP=Placebo              lower       35    NA    NA
#> 2 TRTP=Placebo              quantile    70    NA    NA
#> 3 TRTP=Placebo              upper      177    NA    NA
#> 4 TRTP=Xanomeline High Dose lower        5    25    50
#> 5 TRTP=Xanomeline High Dose quantile    14    36    58
#> 6 TRTP=Xanomeline High Dose upper       22    47    94
#> 7 TRTP=Xanomeline Low Dose  lower       15    28    57
#> 8 TRTP=Xanomeline Low Dose  quantile    19    33    80
#> 9 TRTP=Xanomeline Low Dose  upper       27    51   126

quantile(survfit_object_trt)

#> $quantile
#>                           25 50 75
#> TRTP=Placebo              70 NA NA
#> TRTP=Xanomeline High Dose 14 36 58
#> TRTP=Xanomeline Low Dose  19 33 80
#> 
#> $lower
#>                           25 50 75
#> TRTP=Placebo              35 NA NA
#> TRTP=Xanomeline High Dose  5 25 50
#> TRTP=Xanomeline Low Dose  15 28 57
#> 
#> $upper
#>                            25 50  75
#> TRTP=Placebo              177 NA  NA
#> TRTP=Xanomeline High Dose  22 47  94
#> TRTP=Xanomeline Low Dose   27 51 126
#> 

# }