Obtain risk tables for tables and plots

Create a risktable from a survival object to display in tables or plots

get_risktable(
  survfit_object,
  min_at_risk = 0,
  breaks = NULL,
  statlist = c("n.risk"),
  label = "At risk",
  group = "strata",
  collapse = FALSE
)

# S3 method for survfit
get_risktable(
  survfit_object,
  min_at_risk = 0,
  breaks = NULL,
  statlist = c("n.risk"),
  label = NULL,
  group = "strata",
  collapse = FALSE
)

Arguments

survfit_object
an object of class survfit

min_at_risk
numeric The cutoff for number of subjects to display. Default is 0.

breaks
Single numeric or numeric vector indicating breaks.

statlist
Character vector indicating which summary data to present. Current choices are "n.risk" "n.event" "n.censor".

label
Character vector with labels for the statlist. Default matches "n.risk" with "At risk", "n.event" with "Events" and "n.censor" with "Censored".

group
String indicating the grouping variable for the risk tables. Current options are:

"strata": groups the risk tables per stratum. The label specifies the lables used within each risk table. This is the default

"statlist": groups the risk tables per statlist. The label specifies the title for each risk tabel. The strata levels are used for labeling within each risk table.

"strata" to group the risk tables per strata, and "statlist" to group the risk tables

collapse
Boolean, indicates whether to present the data overall, rather than per strata. Default is FALSE.