Summarize the descriptive statistics across strata from a survival object using S3 method

S3 method for extracting descriptive statistics across strata. No default method is available at the moment.

get_summary(x, ...)

# S3 method for survfit
get_summary(
  survfit_object,
  statlist = c("strata", "records", "events", "median", "LCL", "UCL", "CI"),
  ...
)

Arguments

x
S3 object

...
other arguments passed on to the method

survfit_object
An object of class survfit

statlist
Character vector containing the desired information to be displayed. The order of the arguments determines the order in which they are displayed in the final result. Default is the strata ("strata"), number of subjects ("records"), number of events ("events"), the median survival time ("median"), the Confidence Interval ("CI"), the Lower Confidence Limit ("UCL") and the Upper Confidence Limit ("UCL").

Value

A tibble with summary measures for the Overall Survival Across Strata

Author

Pelagia Alexandra Papadopoulou

Examples

library(survival) 
library(dplyr)
library(tidyr)

survfit_object <- survival::survfit(data = adtte, Surv(AVAL, 1-CNSR) ~ TRTP)
get_summary(survfit_object)

#>                      strata No. of subjects No. of events Median(surv.time)
#> 1              TRTP=Placebo              86            29                NA
#> 2 TRTP=Xanomeline High Dose              84            61                36
#> 3  TRTP=Xanomeline Low Dose              84            62                33
#>   0.95LCL 0.95UCL  0.95CI
#> 1      NA      NA (NA;NA)
#> 2      25      47 (25;47)
#> 3      28      51 (28;51)