Display a summary Table (i.e. table one)

Wrapper function to produce a summary table (i.e. Table One). Create and render a summary table for a dataset. A typical example of a summary table are "table one", the first table in an applied medical research manuscript.

Calculate summary statistics and present them in a formatted table

tableone(
  data,
  title,
  datasource,
  strata = NULL,
  overall = T,
  summary_function = summarize_tab1,
  ...
)

Arguments

data
The dataframe or tibble to visualize

title
Table title to include in the rendered table. Input is a text string.

datasource
String specifying the datasource underlying the data set

strata
Character vector with column names to use for stratification in the summary table. Default: NULL , which indicates no stratification.

overall
If TRUE, the summary statistics for the overall dataset are also calculated

summary_function
A function defining summary statistics for numeric and categorical values Pre-implemented functions are summarize and summarize_tab1

...
Pass options to render_table

footnote
Table footnote to include in the rendered table. Input is a text string.

Examples

library(dplyr)
library(visR)

# metadata for table
t1_title <- "My table one title"
t1_ds <- "My table one data source"
t1_fn <- "My table one footnote"

## table by treatment - without overall and render with DT 
adtte %>%
   dplyr::filter(SAFFL == "Y") %>%
   dplyr::select(AGE, AGEGR1, SEX, EVNTDESC, TRTA) %>%
   visR::tableone(
      strata = "TRTA",
      overall = F,
      title = t1_title,
      datasource = t1_ds,
      footnote = t1_fn,
      engine = "DT"
   )

## table by treatment - without overall and render with GT 
adtte %>%
   dplyr::filter(SAFFL == "Y") %>%
   dplyr::select(AGE, AGEGR1, SEX, EVNTDESC, TRTA) %>%
   visR::tableone(
      strata = "TRTA",
      overall = F,
      title = t1_title,
      datasource = t1_ds,
      footnote = t1_fn,
      engine = "gt"
   )

#> <style>html {
#>   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Helvetica Neue', 'Fira Sans', 'Droid Sans', Arial, sans-serif;
#> }
#> 
#> #mwlefoodfi .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 16px;
#>   font-weight: normal;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #mwlefoodfi .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: center;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #mwlefoodfi .gt_title {
#>   color: #333333;
#>   font-size: 125%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #mwlefoodfi .gt_subtitle {
#>   color: #333333;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 0;
#>   padding-bottom: 4px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #mwlefoodfi .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #mwlefoodfi .gt_col_headings {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #mwlefoodfi .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #mwlefoodfi .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #mwlefoodfi .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #mwlefoodfi .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #mwlefoodfi .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #mwlefoodfi .gt_group_heading {
#>   padding: 8px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #mwlefoodfi .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #mwlefoodfi .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #mwlefoodfi .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #mwlefoodfi .gt_row {
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #mwlefoodfi .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 12px;
#> }
#> 
#> #mwlefoodfi .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #mwlefoodfi .gt_first_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #mwlefoodfi .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #mwlefoodfi .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #mwlefoodfi .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #mwlefoodfi .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #mwlefoodfi .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #mwlefoodfi .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding: 4px;
#> }
#> 
#> #mwlefoodfi .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #mwlefoodfi .gt_sourcenote {
#>   font-size: 90%;
#>   padding: 4px;
#> }
#> 
#> #mwlefoodfi .gt_left {
#>   text-align: left;
#> }
#> 
#> #mwlefoodfi .gt_center {
#>   text-align: center;
#> }
#> 
#> #mwlefoodfi .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #mwlefoodfi .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #mwlefoodfi .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #mwlefoodfi .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #mwlefoodfi .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #mwlefoodfi .gt_footnote_marks {
#>   font-style: italic;
#>   font-size: 65%;
#> }
#> </style>
#> <div id="mwlefoodfi" style="overflow-x:auto;overflow-y:auto;width:auto;height:auto;"><table class="gt_table">
#>   <thead class="gt_header">
#>     <tr>
#>       <th colspan="4" class="gt_heading gt_title gt_font_normal" style>My table one title</th>
#>     </tr>
#>     <tr>
#>       <th colspan="4" class="gt_heading gt_subtitle gt_font_normal gt_bottom_border" style></th>
#>     </tr>
#>   </thead>
#>   <thead class="gt_col_headings">
#>     <tr>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1"></th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1">Placebo (N=86)</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1">Xanomeline High Dose (N=84)</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1">Xanomeline Low Dose (N=84)</th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr class="gt_group_heading_row">
#>       <td colspan="4" class="gt_group_heading">AGE</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left gt_stub">Mean (SD)</td>
#>       <td class="gt_row gt_left">75.2 (8.59)</td>
#>       <td class="gt_row gt_left">74.4 (7.89)</td>
#>       <td class="gt_row gt_left">75.7 (8.29)</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left gt_stub">Median (IQR)</td>
#>       <td class="gt_row gt_left">76 (69.2-81.8)</td>
#>       <td class="gt_row gt_left">76 (70.8-80)</td>
#>       <td class="gt_row gt_left">77.5 (71-82)</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left gt_stub">Min-max</td>
#>       <td class="gt_row gt_left">52-89</td>
#>       <td class="gt_row gt_left">56-88</td>
#>       <td class="gt_row gt_left">51-88</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left gt_stub">Missing</td>
#>       <td class="gt_row gt_left">0 (0%)</td>
#>       <td class="gt_row gt_left">0 (0%)</td>
#>       <td class="gt_row gt_left">0 (0%)</td>
#>     </tr>
#>     <tr class="gt_group_heading_row">
#>       <td colspan="4" class="gt_group_heading">AGEGR1</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left gt_stub">&lt;65</td>
#>       <td class="gt_row gt_left">14 (16.3%)</td>
#>       <td class="gt_row gt_left">11 (13.1%)</td>
#>       <td class="gt_row gt_left">8 (9.52%)</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left gt_stub">&gt;80</td>
#>       <td class="gt_row gt_left">30 (34.9%)</td>
#>       <td class="gt_row gt_left">18 (21.4%)</td>
#>       <td class="gt_row gt_left">29 (34.52%)</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left gt_stub">65-80</td>
#>       <td class="gt_row gt_left">42 (48.8%)</td>
#>       <td class="gt_row gt_left">55 (65.5%)</td>
#>       <td class="gt_row gt_left">47 (55.95%)</td>
#>     </tr>
#>     <tr class="gt_group_heading_row">
#>       <td colspan="4" class="gt_group_heading">SEX</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left gt_stub">F</td>
#>       <td class="gt_row gt_left">53 (61.6%)</td>
#>       <td class="gt_row gt_left">40 (47.6%)</td>
#>       <td class="gt_row gt_left">50 (59.5%)</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left gt_stub">M</td>
#>       <td class="gt_row gt_left">33 (38.4%)</td>
#>       <td class="gt_row gt_left">44 (52.4%)</td>
#>       <td class="gt_row gt_left">34 (40.5%)</td>
#>     </tr>
#>     <tr class="gt_group_heading_row">
#>       <td colspan="4" class="gt_group_heading">EVNTDESC</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left gt_stub">Dematologic Event Occured</td>
#>       <td class="gt_row gt_left">29 (33.7%)</td>
#>       <td class="gt_row gt_left">61 (72.6%)</td>
#>       <td class="gt_row gt_left">62 (73.8%)</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left gt_stub">Study Completion Date</td>
#>       <td class="gt_row gt_left">57 (66.3%)</td>
#>       <td class="gt_row gt_left">23 (27.4%)</td>
#>       <td class="gt_row gt_left">22 (26.2%)</td>
#>     </tr>
#>   </tbody>
#>   <tfoot class="gt_sourcenotes">
#>     <tr>
#>       <td class="gt_sourcenote" colspan="4">Data Source: My table one data source</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_sourcenote" colspan="4">My table one footnote</td>
#>     </tr>
#>   </tfoot>
#>   
#> </table></div>

## table by treatment - without overall and render with kable
adtte %>%
   dplyr::filter(SAFFL == "Y") %>%
   dplyr::select(AGE, AGEGR1, SEX, EVNTDESC, TRTA) %>%
   visR::tableone(
      strata = "TRTA",
      overall = F,
      title = t1_title,
      datasource = t1_ds,
      footnote = t1_fn,
      engine = "kable"
   )

#> <table style="NAborder-bottom: 0;border-bottom: 0;">
#> <caption>My table one title</caption>
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;"> variable </th>
#>    <th style="text-align:left;"> statistic </th>
#>    <th style="text-align:left;"> Placebo (N=86) </th>
#>    <th style="text-align:left;"> Xanomeline High Dose (N=84) </th>
#>    <th style="text-align:left;"> Xanomeline Low Dose (N=84) </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> AGE </td>
#>    <td style="text-align:left;"> Mean (SD) </td>
#>    <td style="text-align:left;"> 75.2 (8.59) </td>
#>    <td style="text-align:left;"> 74.4 (7.89) </td>
#>    <td style="text-align:left;"> 75.7 (8.29) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> AGE </td>
#>    <td style="text-align:left;"> Median (IQR) </td>
#>    <td style="text-align:left;"> 76 (69.2-81.8) </td>
#>    <td style="text-align:left;"> 76 (70.8-80) </td>
#>    <td style="text-align:left;"> 77.5 (71-82) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> AGE </td>
#>    <td style="text-align:left;"> Min-max </td>
#>    <td style="text-align:left;"> 52-89 </td>
#>    <td style="text-align:left;"> 56-88 </td>
#>    <td style="text-align:left;"> 51-88 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> AGE </td>
#>    <td style="text-align:left;"> Missing </td>
#>    <td style="text-align:left;"> 0 (0%) </td>
#>    <td style="text-align:left;"> 0 (0%) </td>
#>    <td style="text-align:left;"> 0 (0%) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> AGEGR1 </td>
#>    <td style="text-align:left;"> &lt;65 </td>
#>    <td style="text-align:left;"> 14 (16.3%) </td>
#>    <td style="text-align:left;"> 11 (13.1%) </td>
#>    <td style="text-align:left;"> 8 (9.52%) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> AGEGR1 </td>
#>    <td style="text-align:left;"> &gt;80 </td>
#>    <td style="text-align:left;"> 30 (34.9%) </td>
#>    <td style="text-align:left;"> 18 (21.4%) </td>
#>    <td style="text-align:left;"> 29 (34.52%) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> AGEGR1 </td>
#>    <td style="text-align:left;"> 65-80 </td>
#>    <td style="text-align:left;"> 42 (48.8%) </td>
#>    <td style="text-align:left;"> 55 (65.5%) </td>
#>    <td style="text-align:left;"> 47 (55.95%) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> SEX </td>
#>    <td style="text-align:left;"> F </td>
#>    <td style="text-align:left;"> 53 (61.6%) </td>
#>    <td style="text-align:left;"> 40 (47.6%) </td>
#>    <td style="text-align:left;"> 50 (59.5%) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> SEX </td>
#>    <td style="text-align:left;"> M </td>
#>    <td style="text-align:left;"> 33 (38.4%) </td>
#>    <td style="text-align:left;"> 44 (52.4%) </td>
#>    <td style="text-align:left;"> 34 (40.5%) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> EVNTDESC </td>
#>    <td style="text-align:left;"> Dematologic Event Occured </td>
#>    <td style="text-align:left;"> 29 (33.7%) </td>
#>    <td style="text-align:left;"> 61 (72.6%) </td>
#>    <td style="text-align:left;"> 62 (73.8%) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> EVNTDESC </td>
#>    <td style="text-align:left;"> Study Completion Date </td>
#>    <td style="text-align:left;"> 57 (66.3%) </td>
#>    <td style="text-align:left;"> 23 (27.4%) </td>
#>    <td style="text-align:left;"> 22 (26.2%) </td>
#>   </tr>
#> </tbody>
#> <tfoot>
#> <tr><td style="padding: 0; " colspan="100%"><span style="font-style: italic;">Additional Note:</span></td></tr>
#> <tr><td style="padding: 0; " colspan="100%">
#> <sup></sup> My table one footnote</td></tr>
#> </tfoot>
#> <tfoot>
#> <tr><td style="padding: 0; " colspan="100%"><span style="font-style: italic;">Data Source:</span></td></tr>
#> <tr><td style="padding: 0; " colspan="100%">
#> <sup></sup> My table one data source</td></tr>
#> </tfoot>
#> </table>

## table by treatment - without overall and render with kable as 
## a latex table format rather than html 
adtte %>%
   dplyr::filter(SAFFL == "Y") %>%
   dplyr::select(AGE, AGEGR1, SEX, EVNTDESC, TRTA) %>%
   visR::tableone(
      strata = "TRTA",
      overall = F,
      title = t1_title,
      datasource = t1_ds,
      footnote = t1_fn,
      output_format = "latex",
      engine = "kable"
   )

#> \begin{table}
#> 
#> \caption{My table one title}
#> \centering
#> \begin{tabular}[t]{lllll}
#> \toprule{}
#> variable & statistic & Placebo (N=86) & Xanomeline High Dose (N=84) & Xanomeline Low Dose (N=84)\\
#> \midrule{}
#>  & Mean (SD) & 75.2 (8.59) & 74.4 (7.89) & 75.7 (8.29)\\
#> \cmidrule{2-5}
#>  & Median (IQR) & 76 (69.2-81.8) & 76 (70.8-80) & 77.5 (71-82)\\
#> \cmidrule{2-5}
#>  & Min-max & 52-89 & 56-88 & 51-88\\
#> \cmidrule{2-5}
#> \multirow[t]{-4}{*}{\raggedright\arraybackslash AGE} & Missing & 0 (0\%) & 0 (0\%) & 0 (0\%)\\
#> \cmidrule{1-5}
#>  & <65 & 14 (16.3\%) & 11 (13.1\%) & 8 (9.52\%)\\
#> \cmidrule{2-5}
#>  & >80 & 30 (34.9\%) & 18 (21.4\%) & 29 (34.52\%)\\
#> \cmidrule{2-5}
#> \multirow[t]{-3}{*}{\raggedright\arraybackslash AGEGR1} & 65-80 & 42 (48.8\%) & 55 (65.5\%) & 47 (55.95\%)\\
#> \cmidrule{1-5}
#>  & F & 53 (61.6\%) & 40 (47.6\%) & 50 (59.5\%)\\
#> \cmidrule{2-5}
#> \multirow[t]{-2}{*}{\raggedright\arraybackslash SEX} & M & 33 (38.4\%) & 44 (52.4\%) & 34 (40.5\%)\\
#> \cmidrule{1-5}
#>  & Dematologic Event Occured & 29 (33.7\%) & 61 (72.6\%) & 62 (73.8\%)\\
#> \cmidrule{2-5}
#> \multirow[t]{-2}{*}{\raggedright\arraybackslash EVNTDESC} & Study Completion Date & 57 (66.3\%) & 23 (27.4\%) & 22 (26.2\%)\\
#> \bottomrule
#> \multicolumn{5}{l}{\rule{0pt}{1em}\textit{Data Source:}}\\
#> \multicolumn{5}{l}{\rule{0pt}{1em}My table one data source}\\
#> \multicolumn{5}{l}{\rule{0pt}{1em}\textit{Additional Note:}}\\
#> \multicolumn{5}{l}{\rule{0pt}{1em}My table one footnote}\\{}
#> \end{tabular}
#> \end{table}