visR: Clinical Graphs and Tables Adhering to Graphical Principles

To enable fit-for-purpose, reusable clinical and medical research focused visualizations and tables with sensible defaults and based on graphical principles as described in: "Vandemeulebroecke et al. (2018)" doi:10.1002/pst.1912 , "Vandemeulebroecke et al. (2019)" doi:10.1002/psp4.12455 , and "Morris et al. (2019)" doi:10.1136/bmjopen-2019-030215 .

See also

Author

Maintainer: Mark Baillie bailliem@gmail.com [copyright holder]

Authors:

Diego Saldana diego.saldana@roche.com

Charlotta Fruechtenicht charlotta.fruechtenicht@roche.com

Marc Vandemeulebroecke marc.vandemeulebroecke@novartis.com

Thanos Siadimas thanos.siadimas@roche.com

Pawel Kawski pawel.kawski@contractors.roche.com

Steven Haesendonckx shaesen2@its.jnj.com

James Black james.black.jb2@roche.com

Pelagia Alexandra Papadopoulou PPapadop@its.jnj.com

Tim Treis tim.treis@outlook.de

Rebecca Albrecht rebecca.albrecht.dietsch@gmail.com

Daniel D. Sjoberg danield.sjoberg@gmail.com (ORCID)

Other contributors:

Ardalan Mirshani ardeeshany@gmail.com [contributor]