Function to summarise risk by interval See code from here https://mcfromnz.wordpress.com/2012/05/05/kaplan-meier-survival-plot-with-at-risk-table-by-sub-groups/

Function to summarise risk by interval See code from here https://mcfromnz.wordpress.com/2012/05/05/kaplan-meier-survival-plot-with-at-risk-table-by-sub-groups/

visr_summary_surv(x, timeby)

Arguments

x
surv object to summarise

timeby
intervals to disaply the risk table

Value

tidy tibble

Examples

library(survival)
library(magrittr)
library(dplyr)
library(ggplot2)
library(broom)
fit <- survival::survfit(Surv(time, status) ~ sex, data = lung)
visr_summary_surv(fit, 50)

#> # A tibble: 42 x 3
#>    strata  time n.risk
#>    <fct>  <dbl>  <dbl>
#>  1 sex=1      0    138
#>  2 sex=1     50    128
#>  3 sex=1    100    114
#>  4 sex=1    150    102
#>  5 sex=1    200     78
#>  6 sex=1    250     62
#>  7 sex=1    300     49
#>  8 sex=1    350     39
#>  9 sex=1    400     31
#> 10 sex=1    450     25
#> # … with 32 more rows