Wrapper for Kaplan Meier analysis for an ADaM Basic Data Structure (BDS) for Time-to-Event analysis

This function is a wrapper around survival::survfit.formula to perform a Kaplan-Meier analysis, based on the expected ADaM Basic Data Structure (BDS) for Time-to-Event analysis and assuming right-censored data. The function expects that the data has been filtered on the PARAM/PARAMCD of interest. Alternatively, PARAM/PARAMCD can be used in the strata argument.
The result is an object of class survfit which can be used in downstream functions and methods that rely on the survfit class. By default:

The Kaplan Meier estimate is estimated directly (stype = 1).

The cumulative hazard is estimated using the Nelson-Aalen estimator (ctype = 1): H.tilde = cumsum(x$n.event/x$n.risk). The MLE (H.hat(t) = -log(S.hat(t))) can't be requested.

A two-sided pointwise 0.95 confidence interval is estimated using a log transformation (conf.type = "log").

vr_KM_est(data = NULL, strata = NULL, ...)

Arguments

data
The name of the ADaM Basic Data Structure (BDS) for Time-to-Event analysis eg ADTTE. Rows in which AVAL or CNSR contain NA, are removed during analysis.

strata
Character vector, representing the strata for Time-to-Event analysis eg TRT01P. When NULL, an overall analysis is performed. Default is NULL.

...
additional arguments passed on to the ellipsis of the call survival::survfit.formula(data = data, formula = Surv(AVAL, 1-CNSR) ~ strata), ...) . Use ?survival::survfit.formula and ?survival::survfitCI for more information.

Value

survfit object, extended by elements PARAM/PARAMCD, ready for downstream processing in estimation or visualization functions and methods.

References

https://https://github.com/therneau/survival

See also

Examples

library(survival)
library(dplyr)
library(tidyr)

## No stratification
vr_KM_est(data = adtte)

#> Call: survfit(formula = Surv(AVAL, 1 - CNSR) ~ 1, data = adtte)
#> 
#>       n  events  median 0.95LCL 0.95UCL 
#>     254     152      51      44      70 

## Stratified Kaplan-Meier analysis by `TRTP`
vr_KM_est(data = adtte, strata = "TRTP")

#> Call: survfit(formula = Surv(AVAL, 1 - CNSR) ~ TRTP, data = adtte)
#> 
#>                            n events median 0.95LCL 0.95UCL
#> TRTP=Placebo              86     29     NA      NA      NA
#> TRTP=Xanomeline High Dose 84     61     36      25      47
#> TRTP=Xanomeline Low Dose  84     62     33      28      51

## Stratified Kaplan-Meier analysis by `TRTP` and `SEX` 
vr_KM_est(data = adtte, strata = c("TRTP", "SEX"))

#> Call: survfit(formula = Surv(AVAL, 1 - CNSR) ~ TRTP + SEX, data = adtte)
#> 
#>                                   n events median 0.95LCL 0.95UCL
#> TRTP=Placebo, SEX=F              53     19     NA      90      NA
#> TRTP=Placebo, SEX=M              33     10     NA      NA      NA
#> TRTP=Xanomeline High Dose, SEX=F 40     27     46      30      70
#> TRTP=Xanomeline High Dose, SEX=M 44     34     25      20      46
#> TRTP=Xanomeline Low Dose, SEX=F  50     34     43      29     100
#> TRTP=Xanomeline Low Dose, SEX=M  34     28     27      21      51

## Stratification with one level
vr_KM_est(data = adtte, strata = "PARAMCD")

#> Call: survfit(formula = Surv(AVAL, 1 - CNSR) ~ PARAMCD, data = adtte)
#> 
#>       n  events  median 0.95LCL 0.95UCL 
#>     254     152      51      44      70 

## Analysis on subset of adtte
vr_KM_est(data = adtte[adtte$SEX == "F", ])

#> Call: survfit(formula = Surv(AVAL, 1 - CNSR) ~ 1, data = adtte[adtte$SEX == 
#>     "F", ])
#> 
#>       n  events  median 0.95LCL 0.95UCL 
#>     143      80      64      47      96 

## Modify the default analysis by using the ellipsis
#vr_KM_est(data = adtte, strata = NULL, type = "kaplan-meier", conf.int = F, timefix = TRUE)
vr_KM_est(data = adtte, strata = NULL, type = "kaplan-meier", conf.int = FALSE, timefix = TRUE)

#> Call: survfit(formula = Surv(AVAL, 1 - CNSR) ~ 1, data = adtte, timefix = TRUE, 
#>     type = "kaplan-meier", conf.int = FALSE)
#> 
#>      n events median 
#>    254    152     51