Create Summary Table (also known as Table 1)

Create a summary table of descriptive statistics from a dataframe or tibble.

By default the following summary stats are calculated:

Numeric variables: mean, min, 25th-percentile, median, 75th-percentile, maximum, standard deviation

Factor variables: proportion of each factor level in the overall dataset

Default: number of unique values and number of missing values

vr_create_tableone(
  data,
  groupCols = NULL,
  overall = TRUE,
  summary_function = vr_summarize_tab1
)

Arguments

data
The dataset to summarize as dataframe or tibble

groupCols
Stratifying/Grouping variable name(s) as character vector. If NULL, only overall results are returned

overall
If TRUE, the summary statistics for the overall dataset are also calculated

summary_function
A function defining summary statistics for numeric and categorical values

Details

It is possible to provide your own summary function. Please have a loot at vr_summary for inspiration.

Note

All columns in the table will be summarized. If only some columns shall be used, please select only those variables prior to creating the summary table by using dplyr::select()

Examples

library(survival)

#> Warning: package 'survival' was built under R version 3.6.3

library(dplyr)

#> Warning: package 'dplyr' was built under R version 3.6.3

#> 
#> Attaching package: 'dplyr'

#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag

#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

ovarian %>%
select(-fustat) %>%
  mutate(age_group = factor(case_when(age <= 50 ~ "<= 50 years",
                                      age <= 60 ~ "<= 60 years",
                                      age <= 70 ~ "<= 70 years",
                                      TRUE ~ "> 70 years")),
         rx = factor(rx),
         ecog.ps = factor(ecog.ps)) %>%
  select(age, age_group, everything()) %>%
  vr_create_tableone()

#> # A tibble: 21 x 3
#>    variable  statistic    Overall                   
#>    <chr>     <chr>        <chr>                     
#>  1 Sample    N            26                        
#>  2 age       Mean (SD)    56.2 (10.1)               
#>  3 age       Median (IQR) 56.8466 (50.16712-62.3781)
#>  4 age       Min-max      38.8932-74.5041           
#>  5 age       Missing      0 (0%)                    
#>  6 age_group <= 50 years  6 (23.1%)                 
#>  7 age_group <= 60 years  13 (50.0%)                
#>  8 age_group <= 70 years  4 (15.4%)                 
#>  9 age_group > 70 years   3 (11.5%)                 
#> 10 futime    Mean (SD)    600 (340)                 
#> # ... with 11 more rows