Render a simple forest plot given a tibble

Render a simple forest plot given a tibble

vr_plot_forest(td)

Arguments

td
tidied tibble

Value

ggplot object

Examples

library(RBesT)

#> Warning: package 'RBesT' was built under R version 3.6.3

#> Loading required package: Rcpp

#> This is RBesT version 1.5.4

#> 
#> Attaching package: 'RBesT'

#> The following object is masked from 'package:survival':
#> 
#>     transplant

library(dplyr)
library(ggplot2)

example(crohn)

#> 
#> crohn> ## Setting up dummy sampling for fast execution of example
#> crohn> ## Please use 4 chains and 20x more warmup & iter in practice
#> crohn> .user_mc_options <- options(RBesT.MC.warmup=50, RBesT.MC.iter=100,
#> crohn+                             RBesT.MC.chains=2, RBesT.MC.thin=1)
#> 
#> crohn> set.seed(546346)
#> 
#> crohn> map_crohn <- gMAP(cbind(y, y.se) ~ 1 | study,
#> crohn+                   family=gaussian,
#> crohn+                   data=transform(crohn, y.se=88/sqrt(n)),
#> crohn+                   weights=n,
#> crohn+                   tau.dist="HalfNormal", tau.prior=44,
#> crohn+                   beta.prior=cbind(0,88))

#> Warning: The largest R-hat is 1.14, indicating chains have not mixed.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#r-hat

#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess

#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#tail-ess

#> Warning: Maximal Rhat > 1.1. Consider increasing RBesT.MC.warmup MCMC parameter.

#> Final MCMC sample equivalent to less than 1000 independent draws.
#> Please consider increasing the MCMC simulation size.

#> 
#> crohn> ## Recover user set sampling defaults
#> crohn> options(.user_mc_options)

vr_tidy_rbest(map_crohn)

#> Error in vr_tidy_rbest(map_crohn): could not find function "vr_tidy_rbest"

map_crohn %>% vr_tidy_rbest() %>% filter(model == "meta") %>%
vr_plot_forest()

#> Error in vr_tidy_rbest(.): could not find function "vr_tidy_rbest"

map_crohn %>% vr_tidy_rbest() %>% filter(model == "stratified") %>%
vr_plot_forest()

#> Error in vr_tidy_rbest(.): could not find function "vr_tidy_rbest"

map_crohn %>% vr_tidy_rbest() %>% vr_plot_forest() + facet_wrap(~ model)

#> Error in vr_tidy_rbest(.): could not find function "vr_tidy_rbest"