Returns a tidied RBesT gMAP object

Returns a tidied RBesT gMAP object

vr_tidy_rbest(x, prob = 0.95)

Arguments

x
RBesT gMAP object

prob
probability range for uncertainty interval

Value

td tidied tibble

Examples

library(RBesT)
example(crohn)

#> 
#> crohn> ## Setting up dummy sampling for fast execution of example
#> crohn> ## Please use 4 chains and 20x more warmup & iter in practice
#> crohn> .user_mc_options <- options(RBesT.MC.warmup=50, RBesT.MC.iter=100,
#> crohn+                             RBesT.MC.chains=2, RBesT.MC.thin=1)
#> 
#> crohn> set.seed(546346)
#> 
#> crohn> map_crohn <- gMAP(cbind(y, y.se) ~ 1 | study,
#> crohn+                   family=gaussian,
#> crohn+                   data=transform(crohn, y.se=88/sqrt(n)),
#> crohn+                   weights=n,
#> crohn+                   tau.dist="HalfNormal", tau.prior=44,
#> crohn+                   beta.prior=cbind(0,88))

#> Warning: The largest R-hat is 1.14, indicating chains have not mixed.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#r-hat

#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess

#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#tail-ess

#> Warning: Maximal Rhat > 1.1. Consider increasing RBesT.MC.warmup MCMC parameter.

#> Final MCMC sample equivalent to less than 1000 independent draws.
#> Please consider increasing the MCMC simulation size.

#> 
#> crohn> ## Recover user set sampling defaults
#> crohn> options(.user_mc_options)

vr_tidy_rbest(map_crohn)

#> # A tibble: 14 x 8
#>    study    estimate    se conf.low conf.high model      study_id row_id
#>    <chr>       <dbl> <dbl>    <dbl>     <dbl> <chr>         <int>  <int>
#>  1 Gastr06     -51   10.2     -71.1    -30.9  stratified        5      1
#>  2 AIMed07     -49    6.83    -62.4    -35.6  stratified        1      2
#>  3 NEJM07      -36    4.86    -45.5    -26.5  stratified        8      3
#>  4 Gastr01a    -47   19.7     -85.6     -8.43 stratified        3      4
#>  5 APhTh04     -90   17.6    -124.     -55.5  stratified        2      5
#>  6 Gastr01b    -54   11.6     -76.6    -31.4  stratified        4      6
#>  7 Gastr06     -48.6  7.56    -71.1    -30.9  meta              5      7
#>  8 AIMed07     -49.3  5.58    -62.4    -35.6  meta              1      8
#>  9 NEJM07      -39.2  5.34    -45.5    -26.5  meta              8      9
#> 10 Gastr01a    -51.2 13.4     -85.6     -8.43 meta              3     10
#> 11 APhTh04     -59.6 14.1    -124.     -55.5  meta              2     11
#> 12 Gastr01b    -50.1  8.21    -76.6    -31.4  meta              4     12
#> 13 MAP         -50.6 15.1     -84.5    -26.9  meta              6     13
#> 14 Mean        -49.4  6.55    -64.7    -38.2  meta              7     14

map_crohn %>%
vr_tidy_rbest()

#> # A tibble: 14 x 8
#>    study    estimate    se conf.low conf.high model      study_id row_id
#>    <chr>       <dbl> <dbl>    <dbl>     <dbl> <chr>         <int>  <int>
#>  1 Gastr06     -51   10.2     -71.1    -30.9  stratified        5      1
#>  2 AIMed07     -49    6.83    -62.4    -35.6  stratified        1      2
#>  3 NEJM07      -36    4.86    -45.5    -26.5  stratified        8      3
#>  4 Gastr01a    -47   19.7     -85.6     -8.43 stratified        3      4
#>  5 APhTh04     -90   17.6    -124.     -55.5  stratified        2      5
#>  6 Gastr01b    -54   11.6     -76.6    -31.4  stratified        4      6
#>  7 Gastr06     -48.6  7.56    -71.1    -30.9  meta              5      7
#>  8 AIMed07     -49.3  5.58    -62.4    -35.6  meta              1      8
#>  9 NEJM07      -39.2  5.34    -45.5    -26.5  meta              8      9
#> 10 Gastr01a    -51.2 13.4     -85.6     -8.43 meta              3     10
#> 11 APhTh04     -59.6 14.1    -124.     -55.5  meta              2     11
#> 12 Gastr01b    -50.1  8.21    -76.6    -31.4  meta              4     12
#> 13 MAP         -50.6 15.1     -84.5    -26.9  meta              6     13
#> 14 Mean        -49.4  6.55    -64.7    -38.2  meta              7     14