Wrapper for Kaplan Meier Time-to-Event analysis

This function is a wrapper around survival::survfit.formula to perform a Kaplan-Meier analysis, assuming right-censored data. The function expects that the data has been filtered on the parameter (PARAM/PARAMCD) of interest. Alternatively, PARAM/PARAMCD can be used in the strata argument.
The result is an object of class survfit which can be used in downstream functions and methods that rely on the survfit class. By default:

The Kaplan Meier estimate is estimated directly (stype = 1).

The cumulative hazard is estimated using the Nelson-Aalen estimator (ctype = 1): H.tilde = cumsum(x$n.event/x$n.risk). The MLE (H.hat(t) = -log(S.hat(t))) can't be requested.

A two-sided pointwise 0.95 confidence interval is estimated using a log transformation (conf.type = "log").

estimate_KM(data = NULL, strata = NULL, CNSR = "CNSR", AVAL = "AVAL", ...)

Arguments

data
The name of the dataset for Time-to-Event analysis based on the Analysis Data Model (ADaM) principles. The dataset is expected to have one record per subject per analysis parameter. Rows in which the analysis variable (AVAL) or the sensor variable (CNSR) contain NA, are removed during analysis.

strata
Character vector, representing the strata for Time-to-Event analysis. When NULL, an overall analysis is performed. Default is NULL.

CNSR
Censor for Time-to-Event analysis. Default is "CNSR", as per CDISC ADaM guiding principles.

AVAL
Analysis value for Time-to-Event analysis. Default is "AVAL", as per CDISC ADaM guiding principles.

...
additional arguments passed on to the ellipsis of the call survival::survfit.formula(data = data, formula = Surv(AVAL, 1-CNSR) ~ strata), ...) . Use ?survival::survfit.formula and ?survival::survfitCI for more information.

Value

survfit object, extended by elements PARAM/PARAMCD, ready for downstream processing in estimation or visualization functions and methods.

References

https://https://github.com/therneau/survival

See also

Author

Steven Haesendonckx

Examples

library(visR)

## No stratification
visR::estimate_KM(data = adtte)

#> Call: survfit(formula = survival::Surv(AVAL, 1 - CNSR) ~ 1, data = adtte)
#> 
#>       n  events  median 0.95LCL 0.95UCL 
#>     254     152      51      44      70 

## Stratified Kaplan-Meier analysis by `TRTP`
visR::estimate_KM(data = adtte, strata = "TRTP")

#> Call: survfit(formula = survival::Surv(AVAL, 1 - CNSR) ~ TRTP, data = adtte)
#> 
#>                            n events median 0.95LCL 0.95UCL
#> TRTP=Placebo              86     29     NA      NA      NA
#> TRTP=Xanomeline High Dose 84     61     36      25      47
#> TRTP=Xanomeline Low Dose  84     62     33      28      51

## Stratified Kaplan-Meier analysis by `TRTP` and `SEX`
visR::estimate_KM(data = adtte, strata = c("TRTP", "SEX"))

#> Call: survfit(formula = survival::Surv(AVAL, 1 - CNSR) ~ TRTP + SEX, 
#>     data = adtte)
#> 
#>                                   n events median 0.95LCL 0.95UCL
#> TRTP=Placebo, SEX=F              53     19     NA      90      NA
#> TRTP=Placebo, SEX=M              33     10     NA      NA      NA
#> TRTP=Xanomeline High Dose, SEX=F 40     27     46      30      70
#> TRTP=Xanomeline High Dose, SEX=M 44     34     25      20      46
#> TRTP=Xanomeline Low Dose, SEX=F  50     34     43      29     100
#> TRTP=Xanomeline Low Dose, SEX=M  34     28     27      21      51

## Stratification with one level
visR::estimate_KM(data = adtte, strata = "PARAMCD")

#> Call: survfit(formula = survival::Surv(AVAL, 1 - CNSR) ~ PARAMCD, data = adtte)
#> 
#>       n  events  median 0.95LCL 0.95UCL 
#>     254     152      51      44      70 

## Analysis on subset of adtte
visR::estimate_KM(data = adtte[adtte$SEX == "F", ])

#> Call: survfit(formula = survival::Surv(AVAL, 1 - CNSR) ~ 1, data = adtte[adtte$SEX == 
#>     "F", ])
#> 
#>       n  events  median 0.95LCL 0.95UCL 
#>     143      80      64      47      96 

## Modify the default analysis by using the ellipsis
estimate_KM(data = adtte, strata = NULL, type = "kaplan-meier", conf.int = F, timefix = TRUE)

#> Call: survfit(formula = survival::Surv(AVAL, 1 - CNSR) ~ 1, data = adtte, 
#>     timefix = TRUE, type = "kaplan-meier", conf.int = FALSE)
#> 
#>      n events median 
#>    254    152     51 

estimate_KM(data = adtte, strata = NULL, type = "kaplan-meier", conf.int = FALSE, timefix = TRUE)

#> Call: survfit(formula = survival::Surv(AVAL, 1 - CNSR) ~ 1, data = adtte, 
#>     timefix = TRUE, type = "kaplan-meier", conf.int = FALSE)
#> 
#>      n events median 
#>    254    152     51 

## Example working with non CDISC data
head(veteran)

#>   trt celltype time status karno diagtime age prior
#> 1   1 squamous   72      1    60        7  69     0
#> 2   1 squamous  411      1    70        5  64    10
#> 3   1 squamous  228      1    60        3  38     0
#> 4   1 squamous  126      1    60        9  63    10
#> 5   1 squamous  118      1    70       11  65    10
#> 6   1 squamous   10      1    20        5  49     0

# convert time and censoring data to ADaM variables 
veteran_adam <- veteran %>%
 mutate(
  AVAL = time, 
  CNSR = if_else(status == 1, 0, 1) #convert censoring status to CDISC principles
  )

estimate_KM(data = veteran_adam, strata = "trt")

#> Call: survfit(formula = survival::Surv(AVAL, 1 - CNSR) ~ trt, data = veteran_adam)
#> 
#>        n events median 0.95LCL 0.95UCL
#> trt=1 69     64  103.0      59     132
#> trt=2 68     64   52.5      44      95