Summarize Hazard Ratio from a survival object using S3 method

S3 method for extracting information regarding Hazard Ratios. No default method is available at the moment.

get_COX_HR(x, ...)

# S3 method for survfit
get_COX_HR(survfit_object = survfit_object, update_formula = NULL, ...)

Arguments

x
S3 object

...
other arguments passed on to the method survival::coxph

survfit_object
An object of class survfit

update_formula
Template which specifies how to update the formula of the survfit_object update.formula

Value

A tibble with Hazard Ratios

See also

Examples

library(survival)
library(dplyr)
library(tidyr)
library(ggplot2)
library(broom)

#> Warning: package 'broom' was built under R version 3.6.3

## treatment effect
survfit_object_trt <- vr_KM_est(data = adtte, strata = c("TRTP"))
get_COX_HR(survfit_object_trt)

#> tidyme S3 default method (broom::tidy) used.

#> # A tibble: 2 x 7
#>   term                  estimate std.error statistic  p.value conf.low conf.high
#>   <chr>                    <dbl>     <dbl>     <dbl>    <dbl>    <dbl>     <dbl>
#> 1 TRTPXanomeline High ~     1.61     0.233      6.92 4.45e-12    1.16       2.07
#> 2 TRTPXanomeline Low D~     1.42     0.230      6.20 5.71e-10    0.973      1.87

## treatment and gender effect
survfit_object_trt_sex <- vr_KM_est(data = adtte, strata = c("TRTP", "SEX"))
get_COX_HR(survfit_object_trt_sex)

#> tidyme S3 default method (broom::tidy) used.

#> # A tibble: 3 x 7
#>   term                  estimate std.error statistic  p.value conf.low conf.high
#>   <chr>                    <dbl>     <dbl>     <dbl>    <dbl>    <dbl>     <dbl>
#> 1 TRTPXanomeline High ~    1.62      0.234      6.91 4.96e-12   1.16       2.07 
#> 2 TRTPXanomeline Low D~    1.48      0.232      6.37 1.92e-10   1.02       1.93 
#> 3 SEXM                     0.375     0.165      2.27 2.34e- 2   0.0508     0.698

## update formula of KM estimates by treatment to include "SEX" for HR estimation
get_COX_HR(survfit_object_trt, update_formula = ". ~ . + SEX")

#> tidyme S3 default method (broom::tidy) used.

#> # A tibble: 3 x 7
#>   term                  estimate std.error statistic  p.value conf.low conf.high
#>   <chr>                    <dbl>     <dbl>     <dbl>    <dbl>    <dbl>     <dbl>
#> 1 TRTPXanomeline High ~    1.62      0.234      6.91 4.96e-12   1.16       2.07 
#> 2 TRTPXanomeline Low D~    1.48      0.232      6.37 1.92e-10   1.02       1.93 
#> 3 SEXM                     0.375     0.165      2.27 2.34e- 2   0.0508     0.698

## update formula of KM estimates by treatment to include "AGE" for HR estimation with ties considered via the efron method
get_COX_HR(survfit_object_trt, update_formula = ". ~ . + strata(AGE)", ties = "efron")

#> tidyme S3 default method (broom::tidy) used.

#> # A tibble: 2 x 7
#>   term                 estimate std.error statistic   p.value conf.low conf.high
#>   <chr>                   <dbl>     <dbl>     <dbl>     <dbl>    <dbl>     <dbl>
#> 1 TRTPXanomeline High~     1.43     0.269      5.30   1.13e-7    0.901      1.96
#> 2 TRTPXanomeline Low ~     1.15     0.265      4.35   1.38e-5    0.633      1.67