Summarize Hazard Ratio from a survival object using S3 method

S3 method for extracting information regarding Hazard Ratios. No default method is available at the moment.

get_COX_HR(x, ...)

# S3 method for survfit
get_COX_HR(survfit_object = survfit_object, update_formula = NULL, ...)

Arguments

x
S3 object

...
other arguments passed on to the method survival::coxph

survfit_object
An object of class survfit

update_formula
Template which specifies how to update the formula of the survfit_object update.formula

Value

A tibble with Hazard Ratios

See also

Author

Steven Haesendonckx

Examples

library(survival)
library(dplyr)
library(tidyr)
library(ggplot2)
library(broom)

## treatment effect
survfit_object_trt <- vr_KM_est(data = adtte, strata = c("TRTP"))

#> Error in vr_KM_est(data = adtte, strata = c("TRTP")): could not find function "vr_KM_est"

get_COX_HR(survfit_object_trt)

#> Error in get_COX_HR(survfit_object_trt): object 'survfit_object_trt' not found

## treatment and gender effect
survfit_object_trt_sex <- vr_KM_est(data = adtte, strata = c("TRTP", "SEX"))

#> Error in vr_KM_est(data = adtte, strata = c("TRTP", "SEX")): could not find function "vr_KM_est"

get_COX_HR(survfit_object_trt_sex)

#> Error in get_COX_HR(survfit_object_trt_sex): object 'survfit_object_trt_sex' not found

## update formula of KM estimates by treatment to include "SEX" for HR estimation
get_COX_HR(survfit_object_trt, update_formula = ". ~ . + SEX")

#> Error in get_COX_HR(survfit_object_trt, update_formula = ". ~ . + SEX"): object 'survfit_object_trt' not found

## update formula of KM estimates by treatment to include "AGE" for
## HR estimation with ties considered via the efron method
get_COX_HR(survfit_object_trt, update_formula = ". ~ . + strata(AGE)", ties = "efron")

#> Error in get_COX_HR(survfit_object_trt, update_formula = ". ~ . + strata(AGE)",     ties = "efron"): object 'survfit_object_trt' not found