#' Create Summary Table (also known as 'tableone')

S3 method for creating summary tables (tableone).

By default the following summary stats are calculated:

Numeric variables: mean, min, 25th-percentile, median, 75th-percentile, maximum, standard deviation

Factor variables: proportion of each factor level in the overall dataset

Default: number of unique values and number of missing values

get_tableone(
  data,
  strata = NULL,
  overall = TRUE,
  summary_function = summarize_tab1
)

# S3 method for default
get_tableone(
  data,
  strata = NULL,
  overall = TRUE,
  summary_function = summarize_tab1
)

Arguments

data
The dataset to summarize as dataframe or tibble

strata
Stratifying/Grouping variable name(s) as character vector. If NULL, only overall results are returned

overall
If TRUE, the summary statistics for the overall dataset are also calculated

summary_function
A function defining summary statistics for numeric and categorical values

Details

It is possible to provide your own summary function. Please have a loot at summary for inspiration.

Note

All columns in the table will be summarized. If only some columns shall be used, please select only those variables prior to creating the summary table by using dplyr::select()

Examples

library(survival)
library(dplyr)
ovarian %>% 
select(-fustat) %>% 
  mutate(age_group = factor(case_when(age <= 50 ~ "<= 50 years",
                                      age <= 60 ~ "<= 60 years",
                                      age <= 70 ~ "<= 70 years",
                                      TRUE ~ "> 70 years")),
         rx = factor(rx),
         ecog.ps = factor(ecog.ps)) %>% 
  select(age, age_group, everything()) %>% 
  get_tableone()

#> # A tibble: 21 x 3
#>    variable  statistic    Total           
#>    <chr>     <chr>        <chr>           
#>  1 Sample    N            26              
#>  2 age       Mean (SD)    56.2 (10.1)     
#>  3 age       Median (IQR) 56.8 (50.2-62.4)
#>  4 age       Min-max      38.9-74.5       
#>  5 age       Missing      0 (0%)          
#>  6 age_group <= 50 years  6 (23.1%)       
#>  7 age_group <= 60 years  13 (50.0%)      
#>  8 age_group <= 70 years  4 (15.4%)       
#>  9 age_group > 70 years   3 (11.5%)       
#> 10 futime    Mean (SD)    600 (340)       
#> # … with 11 more rows